helper class refactoring
